refactor: 푸터 컴포넌트 분리
