+++ client/src/components/FooterIconList/styled.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const FooterIconList = styled.ul`
+  margin-left: 1.7565rem;
+  gap: 1.6rem;
+  display: flex;
+  align-items: center;
+`;
+
+export const SocialLink = styled.a`
+  display: block;
+`;
