+++ client/src/components/FooterMenuList/index.ts
@@ -1,0 +1,1 @@
+export * from "@/components/FooterMenuList/FooterMenuList";
